backend/requirements.txt
alembic==1.13.1
asyncpg==0.29.0
bcrypt==4.1.2
fastapi==0.109.2
httpx==0.26.0
itsdangerous==2.1.2
jinja2==3.1.3
passlib==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
sqlalchemy==2.0.27
structlog==24.1.0
telegram==20.8
uvicorn==0.27.1
psycopg2-binary==2.9.9
pytest==8.0.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-exporter-otlp==1.23.0
prometheus-client==0.19.0 